Add requested fixes.
