+++ backend/moonstream/admin/subscriptions.py
@@ -46,42 +46,3 @@ def migrate_subscriptions(
         new_resources.append(new_resource)
 
     return new_resources
-
-
-def cli_migrate_subscriptions(args: argparse.Namespace) -> None:
-    """
-    Handler for subscriptions migrate.
-    """
-
-    drop_keys = []
-
-    if args.file is not None:
-        with open(args.file) as migration_json_file:
-            migration_json = json.load(migration_json_file)
-
-        if "match" not in migration_json or "update" not in migration_json:
-            print(
-                "Migration file plan have incorrect format require specified {match:{}, update:{}, drop_keys: Optional}"
-            )
-            return
-
-        match = migration_json["match"]
-        update = migration_json["update"]
-
-        if "drop_keys" in migration_json:
-            drop_keys = migration_json["drop_keys"]
-
-    elif args.match is not None and args.update is not None:
-
-        match = json.loads(args.update)
-
-        update = json.loads(args.match)
-
-        if args.drop_keys is not None:
-            drop_keys = json.loads(args.drop_keys)
-
-    else:
-        print("Specified file or --match and --update is required.")
-        return
-
-    migrate_subscriptions(match=match, update=update, drop_keys=drop_keys)
